+++ src/components/characterCreation/CharacterMovesForm.tsx
@@ -1,9 +1,9 @@
-import React, { FC, useEffect, useRef, useState } from 'react';
+import React, { FC, useCallback, useEffect, useRef, useState } from 'react';
 import { useMutation, useQuery } from '@apollo/client';
 import ReactMarkdown from 'react-markdown';
 import { useHistory } from 'react-router-dom';
 import styled from 'styled-components';
-import { Box, CheckBox, Text } from 'grommet';
+import { Box, CheckBox, Tab, Tabs, Text } from 'grommet';
 
 import Spinner from '../Spinner';
 import { accentColors, ButtonWS, HeadingWS, TextWS } from '../../config/grommetConfig';
@@ -13,6 +13,8 @@ import { CharacterCreationSteps } from '../../@types/enums';
 import { useFonts } from '../../contexts/fontContext';
 import { useGame } from '../../contexts/gameContext';
 import { CharacterMove, Move } from '../../@types/staticDataInterfaces';
+import { decapitalize } from '../../helpers/decapitalize';
+import OTHER_PLAYBOOK_MOVES, { OtherPlaybookMovesData, OtherPlaybookMovesVars } from '../../queries/otherPlaybookMoves';
 
 const StyledMarkdown = styled(ReactMarkdown)`
   & p {
@@ -26,13 +28,14 @@ export const INCREASED_BY_IMPROVEMENT_TEXT = '(increased by improvement)';
 const CharacterMovesForm: FC = () => {
   // -------------------------------------------------- Component state ---------------------------------------------------- //
   const [selectedMoves, setSelectedMoves] = useState<{ id: string; name: string }[]>([]);
+  const [activeTab, setActiveTab] = useState(0);
   let hasInitialised = useRef(false);
   // ------------------------------------------------------- Hooks --------------------------------------------------------- //
   const { game, character, userGameRole } = useGame();
   const { crustReady } = useFonts();
 
   const allowedPlaybookMoves = character?.allowedPlaybookMoves;
-  // const allowedOtherPlaybookMoves = character?.allowedOtherPlaybookMoves;
+  const allowedOtherPlaybookMoves = character?.allowedOtherPlaybookMoves;
 
   // --------------------------------------------------3rd party hooks ----------------------------------------------------- //
   const history = useHistory();
@@ -44,6 +47,13 @@ const CharacterMovesForm: FC = () => {
     { variables: { playbookType: character?.playbook }, skip: !character?.playbook }
   );
 
+  const { data: otherMovesData } = useQuery<OtherPlaybookMovesData, OtherPlaybookMovesVars>(OTHER_PLAYBOOK_MOVES, {
+    // @ts-ignore
+    variables: { playbookType: character?.playbook },
+    skip: !character || character.allowedOtherPlaybookMoves < 1,
+  });
+  const otherPlaybookMoves = otherMovesData?.otherPlaybookMoves;
+
   const optionalMoves = pbCreatorData?.playbookCreator.optionalMoves;
   const defaultMoves = pbCreatorData?.playbookCreator.defaultMoves;
   const moveChoiceCount = pbCreatorData?.playbookCreator.moveChoiceCount;
@@ -53,13 +63,24 @@ const CharacterMovesForm: FC = () => {
   );
 
   // ------------------------------------------ Component functions and variables ------------------------------------------ //
+
+  const getTotalAllowedMoves = (): number | undefined => {
+    if (!!character && !!pbCreatorData) {
+      return (
+        character.allowedPlaybookMoves +
+        pbCreatorData.playbookCreator.defaultMoves.length +
+        character.allowedOtherPlaybookMoves
+      );
+    }
+  };
   const handleSelectMove = (move: Move) => {
+    const totalAllowedMoves = getTotalAllowedMoves();
+
     if (selectedMoves.some((item) => item.name === move.name)) {
       setSelectedMoves(selectedMoves.filter((item) => item.name !== move.name));
     } else {
-      if (!!allowedPlaybookMoves && !!defaultMoves) {
-        selectedMoves.length < allowedPlaybookMoves + defaultMoves.length &&
-          setSelectedMoves([...selectedMoves, { id: move.id, name: move.name }]);
+      if (!!totalAllowedMoves) {
+        selectedMoves.length < totalAllowedMoves && setSelectedMoves([...selectedMoves, { id: move.id, name: move.name }]);
       }
     }
   };
@@ -82,24 +103,106 @@ const CharacterMovesForm: FC = () => {
 
   // ------------------------------------------------------ Effects -------------------------------------------------------- //
 
+  const setMatchingMoves = useCallback(
+    (originalMoves: Move[]) => {
+      !!character &&
+        character.characterMoves.forEach((characterMove: CharacterMove) => {
+          const matchingMove: Move | undefined = originalMoves.find((om) => om.name === characterMove.name);
+          !!matchingMove &&
+            setSelectedMoves((prevState) => [...prevState, { id: matchingMove.id, name: matchingMove.name }]);
+        });
+    },
+    [character]
+  );
+
   // load in existing moves, but prevent duplicates
   useEffect(() => {
     if (!!character && !hasInitialised.current && !!optionalMoves && !!defaultMoves) {
-      character.characterMoves.forEach((characterMove: CharacterMove) => {
-        // The CharacterMoves have a different id, so need the original Move equivalent to the CharacterMove
-        const originalMoves: Move[] = optionalMoves.concat(defaultMoves);
-        const matchingMove: Move | undefined = originalMoves.find((om) => om.name === characterMove.name);
-        !!matchingMove && setSelectedMoves((prevState) => [...prevState, { id: matchingMove.id, name: matchingMove.name }]);
-      });
-      hasInitialised.current = true;
+      // The CharacterMoves have a different id, so need the original Move equivalent to the CharacterMove
+      let originalMoves: Move[];
+      if (!!allowedOtherPlaybookMoves && allowedOtherPlaybookMoves > 0 && !!otherPlaybookMoves) {
+        originalMoves = optionalMoves.concat(defaultMoves).concat(otherPlaybookMoves);
+        setMatchingMoves(originalMoves);
+        hasInitialised.current = true;
+      } else if (allowedOtherPlaybookMoves === 0) {
+        originalMoves = optionalMoves.concat(defaultMoves);
+        setMatchingMoves(originalMoves);
+        hasInitialised.current = true;
+      }
     }
-  }, [character, optionalMoves, defaultMoves]);
+  }, [character, optionalMoves, defaultMoves, allowedOtherPlaybookMoves, otherPlaybookMoves, setMatchingMoves]);
 
-  if (!allowedPlaybookMoves || !defaultMoves) {
+  // -------------------------------------------------- Render component  ---------------------------------------------------- //
+  if (!defaultMoves || !character || !getTotalAllowedMoves()) {
     return <Spinner />;
   }
 
-  // -------------------------------------------------- Render component  ---------------------------------------------------- //
+  const playbookMovesList = (
+    <>
+      {!!moveChoiceCount && !!allowedPlaybookMoves && moveChoiceCount > 0 && (
+        <Box direction="row" align="center" gap="12px">
+          <Text size="large" weight="bold" margin={{ vertical: '12px' }}>
+            {`Select ${moveChoiceCount < allowedPlaybookMoves ? allowedPlaybookMoves : moveChoiceCount}`}
+          </Text>
+          {moveChoiceCount < allowedPlaybookMoves && (
+            <TextWS color={accentColors[0]}>{INCREASED_BY_IMPROVEMENT_TEXT}</TextWS>
+          )}
+        </Box>
+      )}
+      <Box align="start" gap="12px">
+        {!!optionalMoves &&
+          optionalMoves.map((move) => {
+            return (
+              <CheckBox
+                key={move.id}
+                label={
+                  <div>
+                    <StyledMarkdown>{move.description}</StyledMarkdown>
+                  </div>
+                }
+                checked={selectedMoves.some((mv) => mv.name === move.name)}
+                onChange={() => handleSelectMove(move)}
+              />
+            );
+          })}
+      </Box>
+    </>
+  );
+
+  const otherMovesList = (
+    <>
+      <Box>
+        <Text size="large" weight="bold" margin={{ vertical: '12px' }}>
+          {`Select ${allowedOtherPlaybookMoves}`}
+        </Text>
+      </Box>
+      <Box align="start" gap="12px">
+        {!!otherPlaybookMoves &&
+          otherPlaybookMoves.map((move) => {
+            return (
+              <CheckBox
+                key={move.id}
+                label={
+                  <div>
+                    <StyledMarkdown>{move.description}</StyledMarkdown>
+                  </div>
+                }
+                checked={selectedMoves.some((mv) => mv.name === move.name)}
+                onChange={() => handleSelectMove(move)}
+              />
+            );
+          })}
+      </Box>
+    </>
+  );
+
+  const tabbedMoveLists = (
+    <Tabs activeIndex={activeTab} onActive={(tab) => setActiveTab(tab)} margin={{ top: '6px' }}>
+      <Tab title={`${decapitalize(character.playbook)} moves`}>{playbookMovesList}</Tab>
+      <Tab title="Other moves">{otherMovesList}</Tab>
+    </Tabs>
+  );
+
   return (
     <Box
       data-testid="character-moves-form"
@@ -119,7 +222,7 @@ const CharacterMovesForm: FC = () => {
           <ButtonWS
             primary
             label={settingMoves ? <Spinner fillColor="#FFF" width="37px" height="36px" /> : 'SET'}
-            disabled={selectedMoves.length !== allowedPlaybookMoves + defaultMoves.length}
+            disabled={selectedMoves.length !== getTotalAllowedMoves()}
             onClick={() =>
               !settingMoves && handleSubmitCharacterMoves([...selectedMoves.map((mv) => mv.id), ...defaultMoveIds])
             }
@@ -137,33 +240,7 @@ const CharacterMovesForm: FC = () => {
           defaultMoves.map((move) => (
             <CheckBox key={move.id} checked label={<StyledMarkdown>{move.description}</StyledMarkdown>} />
           ))}
-        {!!moveChoiceCount && !!allowedPlaybookMoves && moveChoiceCount > 0 && (
-          <Box direction="row" align="center" gap="12px">
-            <Text size="large" weight="bold" margin={{ vertical: '12px' }}>
-              {`Select ${moveChoiceCount < allowedPlaybookMoves ? allowedPlaybookMoves : moveChoiceCount}`}
-            </Text>
-            {moveChoiceCount < allowedPlaybookMoves && (
-              <TextWS color={accentColors[0]}>{INCREASED_BY_IMPROVEMENT_TEXT}</TextWS>
-            )}
-          </Box>
-        )}
-        <Box align="start" gap="12px">
-          {!!optionalMoves &&
-            optionalMoves.map((move) => {
-              return (
-                <CheckBox
-                  key={move.id}
-                  label={
-                    <div>
-                      <StyledMarkdown>{move.description}</StyledMarkdown>
-                    </div>
-                  }
-                  checked={selectedMoves.some((mv) => mv.name === move.name)}
-                  onChange={() => handleSelectMove(move)}
-                />
-              );
-            })}
-        </Box>
+        {!!allowedOtherPlaybookMoves && allowedOtherPlaybookMoves > 0 ? tabbedMoveLists : playbookMovesList}
       </Box>
     </Box>
   );
