+++ src/components/characterCreation/__tests__/CharacterMovesForm.test.tsx
@@ -1,14 +1,13 @@
 import React from 'react';
 import userEvent from '@testing-library/user-event';
 import { InMemoryCache } from '@apollo/client';
-import { screen } from '@testing-library/react';
+import { getAllByRole, screen } from '@testing-library/react';
 
 import CharacterMovesForm, { INCREASED_BY_IMPROVEMENT_TEXT } from '../CharacterMovesForm';
 import { mockKeycloakStub } from '../../../../__mocks__/@react-keycloak/web';
 import {
   blankCharacter,
   mockCharacter2,
-  mockCharacterMoveAngel1,
   mockGame5,
   mockKeycloakUserInfo1,
   mockPlaybookCreatorAngel,
@@ -16,6 +15,18 @@ import {
 import { renderWithRouter, waitOneTick } from '../../../tests/test-utils';
 import { mockPlaybookCreator } from '../../../tests/mockQueries';
 import { Game } from '../../../@types/dataInterfaces';
+import { CharacterMove } from '../../../@types/staticDataInterfaces';
+import {
+  mockCharacterMoveAngel1,
+  mockCharacterMoveAngel2,
+  mockCharacterMoveAngel3,
+  mockCharacterMoveAngel4,
+  mockSeeingSoulsAsCM,
+} from '../../../tests/mockData/mockCharacterMoves';
+import { MockedResponse } from '@apollo/client/testing';
+import OTHER_PLAYBOOK_MOVES, { OtherPlaybookMovesData } from '../../../queries/otherPlaybookMoves';
+import { mockBonefeel, mockEverybodyEats, mockSeeingSouls } from '../../../tests/mockData/mockMoves';
+import { decapitalize } from '../../../helpers/decapitalize';
 
 jest.mock('@react-keycloak/web', () => {
   const originalModule = jest.requireActual('@react-keycloak/web');
@@ -25,7 +36,26 @@ jest.mock('@react-keycloak/web', () => {
   };
 });
 
-const generateGame = (allowedPlaybookMoves: number): Game => ({
+const mockOtherPlaybookMovesQuery: MockedResponse<OtherPlaybookMovesData> = {
+  request: {
+    query: OTHER_PLAYBOOK_MOVES,
+    variables: { playbookType: mockCharacter2.playbook },
+  },
+  result: () => {
+    // console.log('mockOtherPlaybookMovesQuery');
+    return {
+      data: {
+        otherPlaybookMoves: [mockBonefeel, mockEverybodyEats, mockSeeingSouls],
+      },
+    };
+  },
+};
+
+const generateGame = (
+  allowedPlaybookMoves: number,
+  allowedOtherPlaybookMoves: number,
+  characterMoves: CharacterMove[]
+): Game => ({
   ...mockGame5,
   gameRoles: [
     mockGame5.gameRoles[0],
@@ -49,7 +79,8 @@ const generateGame = (allowedPlaybookMoves: number): Game => ({
           gear: mockCharacter2.gear,
           playbookUniques: mockCharacter2.playbookUniques,
           allowedPlaybookMoves,
-          characterMoves: [{ ...mockCharacterMoveAngel1, isSelected: true }],
+          allowedOtherPlaybookMoves,
+          characterMoves,
         },
       ],
     },
@@ -63,75 +94,215 @@ describe('Rendering CharacterMovesForm', () => {
     cache = new InMemoryCache();
   });
 
-  test('should render CharacterMovesForm in initial state (Angel)', async () => {
-    renderWithRouter(<CharacterMovesForm />, `/character-creation/${mockGame5.id}?step=7`, {
-      isAuthenticated: true,
-      injectedGame: generateGame(2),
-      apolloMocks: [mockPlaybookCreator],
-      injectedUserId: mockKeycloakUserInfo1.sub,
-      cache,
-    });
+  describe('with only default moves on character', () => {
+    beforeEach(async () => {
+      renderWithRouter(<CharacterMovesForm />, `/character-creation/${mockGame5.id}?step=7`, {
+        isAuthenticated: true,
+        injectedGame: generateGame(2, 0, [{ ...mockCharacterMoveAngel1, isSelected: true }]),
+        apolloMocks: [mockPlaybookCreator],
+        injectedUserId: mockKeycloakUserInfo1.sub,
+        cache,
+      });
 
-    await waitOneTick();
-
-    screen.getByTestId('character-moves-form');
-    screen.getByRole('heading', { name: `WHAT ARE ${mockCharacter2.name?.toUpperCase()}'S MOVES?` });
-    const setButton = screen.getByRole('button', { name: 'SET' }) as HTMLButtonElement;
-    expect(setButton.disabled).toEqual(true);
-    const defaultMoveCheckbox = screen.getByRole('checkbox', {
-      name: mockPlaybookCreatorAngel.defaultMoves[0].description,
-    }) as HTMLInputElement;
-    expect(defaultMoveCheckbox.checked).toEqual(true);
-    mockPlaybookCreatorAngel.optionalMoves.forEach((move) => {
-      const checkbox = screen.getByRole('checkbox', { name: move.description }) as HTMLInputElement;
-      expect(checkbox.checked).toEqual(false);
+      await waitOneTick();
     });
-    expect(screen.queryByText(INCREASED_BY_IMPROVEMENT_TEXT)).not.toBeInTheDocument();
+    test('should render CharacterMovesForm in initial state (Angel)', async () => {
+      screen.getByTestId('character-moves-form');
+      screen.getByRole('heading', { name: `WHAT ARE ${mockCharacter2.name?.toUpperCase()}'S MOVES?` });
+      const setButton = screen.getByRole('button', { name: 'SET' }) as HTMLButtonElement;
+      expect(setButton.disabled).toBeTruthy();
+      const defaultMoveCheckbox = screen.getByRole('checkbox', {
+        name: mockPlaybookCreatorAngel.defaultMoves[0].description,
+      }) as HTMLInputElement;
+      expect(defaultMoveCheckbox.checked).toBeTruthy();
+      mockPlaybookCreatorAngel.optionalMoves.forEach((move) => {
+        const checkbox = screen.getByRole('checkbox', { name: move.description }) as HTMLInputElement;
+        expect(checkbox.checked).toBeFalsy();
+      });
+      expect(screen.queryByText(INCREASED_BY_IMPROVEMENT_TEXT)).not.toBeInTheDocument();
+      expect(screen.queryByRole('tab', { name: `${decapitalize(mockCharacter2.playbook)} moves` })).not.toBeInTheDocument();
+      expect(screen.queryByRole('tab', { name: 'Other moves' })).not.toBeInTheDocument();
+    });
+    test('should select 2 moves and enable SET button', async () => {
+      const setButton = screen.getByRole('button', { name: 'SET' }) as HTMLButtonElement;
+      expect(setButton.disabled).toBeTruthy();
+      const checkbox1 = screen.getByRole('checkbox', {
+        name: mockPlaybookCreatorAngel.optionalMoves[0].description,
+      }) as HTMLInputElement;
+      const checkbox2 = screen.getByRole('checkbox', {
+        name: mockPlaybookCreatorAngel.optionalMoves[1].description,
+      }) as HTMLInputElement;
+
+      userEvent.click(checkbox1);
+      expect(checkbox1.checked).toBeTruthy();
+
+      userEvent.click(checkbox2);
+      expect(checkbox2.checked).toBeTruthy();
+
+      expect(setButton.disabled).toBeFalsy();
+    });
   });
 
-  test('should render CharacterMovesForm with increased allowed moves', async () => {
-    renderWithRouter(<CharacterMovesForm />, `/character-creation/${mockGame5.id}?step=7`, {
-      isAuthenticated: true,
-      injectedGame: generateGame(3),
-      apolloMocks: [mockPlaybookCreator],
-      injectedUserId: mockKeycloakUserInfo1.sub,
-      cache,
+  describe('with default moves and 2 optional moves on character', () => {
+    beforeEach(async () => {
+      renderWithRouter(<CharacterMovesForm />, `/character-creation/${mockGame5.id}?step=7`, {
+        isAuthenticated: true,
+        injectedGame: generateGame(2, 0, [
+          { ...mockCharacterMoveAngel1, isSelected: true },
+          { ...mockCharacterMoveAngel2, isSelected: true },
+          { ...mockCharacterMoveAngel3, isSelected: true },
+        ]),
+        apolloMocks: [mockPlaybookCreator],
+        injectedUserId: mockKeycloakUserInfo1.sub,
+        cache,
+      });
+
+      await waitOneTick();
     });
 
-    await waitOneTick();
+    test('should render CharacterMovesForm with correct initial state', () => {
+      expect(screen.getByText('Select 2')).toBeInTheDocument();
+      expect(screen.queryByText(INCREASED_BY_IMPROVEMENT_TEXT)).not.toBeInTheDocument();
+      const checkboxes = screen.getAllByRole('checkbox') as HTMLInputElement[];
+      expect(checkboxes.filter((cb) => cb.checked)).toHaveLength(3);
+    });
 
-    screen.getByTestId('character-moves-form');
-    expect(screen.getByText('Select 3')).toBeInTheDocument();
-    expect(screen.getByText(INCREASED_BY_IMPROVEMENT_TEXT)).toBeInTheDocument();
+    test('should deselect a move and select a different one', () => {
+      const setButton = screen.getByRole('button', { name: 'SET' }) as HTMLButtonElement;
+      const checkboxes = screen.getAllByRole('checkbox') as HTMLInputElement[];
+      const checkedCheckboxes = checkboxes.filter((cb) => cb.checked);
+      const uncheckedCheckboxes = checkboxes.filter((cb) => !cb.checked);
+
+      // Check that user can't unselect DEFAULT_CHARACTER moves
+      userEvent.click(checkedCheckboxes[0]);
+      expect(checkedCheckboxes[0].checked).toBeTruthy();
+
+      // Uncheck an optional move
+      userEvent.click(checkedCheckboxes[1]);
+      expect(checkedCheckboxes[1].checked).toBeFalsy();
+
+      // Select a different optional move
+      userEvent.click(uncheckedCheckboxes[0]);
+      expect(uncheckedCheckboxes[0].checked).toBeTruthy();
+
+      expect(setButton.disabled).toBeFalsy();
+    });
   });
 
-  test('should select 2 moves and enable SET button', async () => {
-    renderWithRouter(<CharacterMovesForm />, `/character-creation/${mockGame5.id}?step=7`, {
-      isAuthenticated: true,
-      injectedGame: generateGame(2),
-      apolloMocks: [mockPlaybookCreator],
-      injectedUserId: mockKeycloakUserInfo1.sub,
-      cache,
+  describe('with default moves, 2 optional moves, and one ADD_CHARACTER_MOVE improvement', () => {
+    beforeEach(async () => {
+      renderWithRouter(<CharacterMovesForm />, `/character-creation/${mockGame5.id}?step=7`, {
+        isAuthenticated: true,
+        injectedGame: generateGame(3, 0, [
+          { ...mockCharacterMoveAngel1, isSelected: true },
+          { ...mockCharacterMoveAngel2, isSelected: true },
+          { ...mockCharacterMoveAngel3, isSelected: true },
+        ]),
+        apolloMocks: [mockPlaybookCreator],
+        injectedUserId: mockKeycloakUserInfo1.sub,
+        cache,
+      });
+
+      await waitOneTick();
+    });
+    test('should render CharacterMovesForm with correct initial state', async () => {
+      screen.getByTestId('character-moves-form');
+      expect(screen.getByText('Select 3')).toBeInTheDocument();
+      expect(screen.getByText(INCREASED_BY_IMPROVEMENT_TEXT)).toBeInTheDocument();
+      const checkboxes = screen.getAllByRole('checkbox') as HTMLInputElement[];
+      expect(checkboxes.filter((cb) => cb.checked)).toHaveLength(3);
     });
 
-    await waitOneTick();
+    test('should select and add 4th CharacterMove', () => {
+      const setButton = screen.getByRole('button', { name: 'SET' }) as HTMLButtonElement;
+      expect(setButton.disabled).toBeTruthy();
+      const checkbox3 = screen.getByRole('checkbox', {
+        name: mockPlaybookCreatorAngel.optionalMoves[2].description,
+      }) as HTMLInputElement;
 
-    screen.getByTestId('character-moves-form');
-    const setButton = screen.getByRole('button', { name: 'SET' }) as HTMLButtonElement;
-    expect(setButton.disabled).toEqual(true);
-    const checkbox1 = screen.getByRole('checkbox', {
-      name: mockPlaybookCreatorAngel.optionalMoves[0].description,
-    }) as HTMLInputElement;
-    const checkbox2 = screen.getByRole('checkbox', {
-      name: mockPlaybookCreatorAngel.optionalMoves[1].description,
-    }) as HTMLInputElement;
+      userEvent.click(checkbox3);
+      expect(checkbox3.checked).toBeTruthy();
 
-    userEvent.click(checkbox1);
-    expect(checkbox1.checked).toEqual(true);
+      expect(setButton.disabled).toBeFalsy();
+      const checkboxes = screen.getAllByRole('checkbox') as HTMLInputElement[];
+      expect(checkboxes.filter((cb) => cb.checked)).toHaveLength(4);
+    });
+  });
 
-    userEvent.click(checkbox2);
-    expect(checkbox2.checked).toEqual(true);
+  describe('with default moves, 3 optional moves, and one ADD_OTHER_PB_MOVE improvement', () => {
+    beforeEach(async () => {
+      renderWithRouter(<CharacterMovesForm />, `/character-creation/${mockGame5.id}?step=7`, {
+        isAuthenticated: true,
+        injectedGame: generateGame(3, 1, [
+          { ...mockCharacterMoveAngel1, isSelected: true },
+          { ...mockCharacterMoveAngel2, isSelected: true },
+          { ...mockCharacterMoveAngel3, isSelected: true },
+          { ...mockCharacterMoveAngel4, isSelected: true },
+        ]),
+        apolloMocks: [mockPlaybookCreator, mockOtherPlaybookMovesQuery],
+        injectedUserId: mockKeycloakUserInfo1.sub,
+        cache,
+      });
 
-    expect(setButton.disabled).toEqual(false);
+      await waitOneTick(); // wait for game query & otherPlaybookMoves query
+    });
+
+    test('should render CharacterMovesForm with correct initial state', () => {
+      expect(screen.getByText('Select 3')).toBeInTheDocument();
+      expect(screen.queryByText(INCREASED_BY_IMPROVEMENT_TEXT)).toBeInTheDocument();
+      const checkboxes = screen.getAllByRole('checkbox') as HTMLInputElement[];
+      expect(checkboxes.filter((cb) => cb.checked)).toHaveLength(4);
+      expect(screen.getByRole('tab', { name: `${decapitalize(mockCharacter2.playbook)} moves` })).toBeInTheDocument();
+      expect(screen.getByRole('tab', { name: 'Other moves' })).toBeInTheDocument();
+    });
+
+    test('should change tab and select move from other playbook', () => {
+      const setButton = screen.getByRole('button', { name: 'SET' }) as HTMLButtonElement;
+      userEvent.click(screen.getByRole('tab', { name: 'Other moves' }));
+
+      expect(screen.getByRole('tabpanel', { name: 'Other moves Tab Contents' })).toBeInTheDocument();
+      const checkboxes = screen.getAllByRole('checkbox') as HTMLInputElement[];
+      expect(checkboxes).toHaveLength(4); // 1 default move + 3 moves from other playbooks
+
+      // Select move from other playbook
+      userEvent.click(checkboxes[1]);
+      expect(checkboxes[1].checked).toBeTruthy();
+
+      // Check SET button enables
+      expect(setButton.disabled).toBeFalsy();
+    });
+  });
+
+  describe('with default moves and 4 moves on character (one from other playbook)', () => {
+    beforeEach(async () => {
+      renderWithRouter(<CharacterMovesForm />, `/character-creation/${mockGame5.id}?step=7`, {
+        isAuthenticated: true,
+        injectedGame: generateGame(3, 1, [
+          { ...mockCharacterMoveAngel1, isSelected: true },
+          { ...mockCharacterMoveAngel2, isSelected: true },
+          { ...mockCharacterMoveAngel3, isSelected: true },
+          { ...mockCharacterMoveAngel4, isSelected: true },
+          { ...mockSeeingSoulsAsCM, isSelected: true },
+        ]),
+        apolloMocks: [mockPlaybookCreator, mockOtherPlaybookMovesQuery],
+        injectedUserId: mockKeycloakUserInfo1.sub,
+        cache,
+      });
+
+      await waitOneTick(); // wait for game query & otherPlaybookMoves query
+    });
+
+    test('should should render CharacterMovesForm with correct initial state', () => {
+      expect(screen.getByText('Select 3')).toBeInTheDocument();
+      expect(screen.queryByText(INCREASED_BY_IMPROVEMENT_TEXT)).toBeInTheDocument();
+      let checkboxes = screen.getAllByRole('checkbox') as HTMLInputElement[];
+      expect(checkboxes.filter((cb) => cb.checked)).toHaveLength(4);
+      expect(screen.getByRole('tab', { name: `${decapitalize(mockCharacter2.playbook)} moves` })).toBeInTheDocument();
+      expect(screen.getByRole('tab', { name: 'Other moves' })).toBeInTheDocument();
+
+      userEvent.click(screen.getByRole('tab', { name: 'Other moves' }));
+      checkboxes = screen.getAllByRole('checkbox') as HTMLInputElement[];
+      expect(checkboxes.filter((cb) => cb.checked)).toHaveLength(2); // 1 selected default move + 1 selected other playbook move
+    });
   });
 });
