+++ src/components/dialogs/BoardVehicleDialog.tsx
@@ -14,7 +14,7 @@ import { CharacterMove, Move } from '../../@types/staticDataInterfaces';
 import { useFonts } from '../../contexts/fontContext';
 import { useGame } from '../../contexts/gameContext';
 import { Vehicle } from '../../@types/dataInterfaces';
-import { dummyVehicleFrame } from '../../tests/mocks';
+import { dummyVehicleFrame } from '../../tests/mockData/dummyData';
 import { VehicleType } from '../../@types/enums';
 
 interface BoardVehicleDialogProps {
